More explicitly require `x86_64_v4` target for root specs in avx512 environments
Similarly to f10f4583051ebe5c977fe974b577e3f8a7734fb3 (in #68),
`packages:all:target` isn't strict enough for forcing the packages to be built
for that target.  At the same time we can't use `packages:all:require` here
because we want to include the `avx2_buildtools` and `multiarch` environments,
which would be incompatible with the requirement of having *all* packages
targeting `x86_64_v4`.  Our solution is to append `target=x86_64_v4` to force
that architecture only for the root specs in the avx512 environments, but allow
dependencies to be reused from the `x86_64_v3` environments.