Resolve ambiguity in module name generation
Spack uses a "first matching" strategy when choosing the name of the module, in
the collective `youngmichael` environment we have both `openmpi +cuda` and
`openmpi ~cuda` specs, but they had conflicting module names because only the
`openmpi` projection was matched, not also the `^cuda` one.  By adding an
explicit `openmpi +cuda` projection before `openmpi` we can resolve the
ambiguity.